refactor(logger): consolidate logger implementation and remove unused files
