--- src/logger/audit.logger.ts
@@ -1,1 +1,0 @@
-//TODO: implement
